Highlight current METAR LED
I wanted to highlight the Map LED of the airport that is currently displaying the METAR on the screen. Thought that would be cool. 

In the future it would be awesome to put some proximity buttons on the frame and scroll through manually or go back if you missed one. Probably won't happen.

I'm like the worst programmer in the world, but I wanted to try to do this to learn. Perhaps you'll delete this, which I'm okay with. I probably could have used less variables. Because I didn't want to upload the whole file, I'm hoping I got everything. It doesn't even start working until the second METAR.

Thanks for giving me this project to explore my limited programming skills. My next project is to make a welcome screen for the display. Many people give these as gifts. So Property of.... Many thanks too....